Finding sites with variation

Single nucleotide polymorphism (SNP) in the whole package refers to variation of amino acid. SNPsite will try to find SNP in the multiple sequence alignment. A reference sequence and gap character may be specified to number the site.

SNPsites(tree, ...)

# S3 method for phyMSAmatched
SNPsites(tree, minSNP = NULL, ...)

Arguments

tree
A phyMSAmatched object.

...
Other arguments

minSNP
Minimum number of a mutation to be a SNP. The default is 10th of the total tree tips.

Value

A SNPsites object.

Examples

data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
SNPsites(tree)
#>  [1]   21   80  105  106  123  130  139  151  196  229  257  313  346  631  633
#> [16]  727  763  777  784  840  894  916  940  956  982  988 1120 1143 1204 1205
#> [31] 1226 1274 1285 1289 1411 1439 1477 1542 1673 1717 1857 1902 1949 2068 2074
#> [46] 2086 2197 2261 2284 2295 2367 2445 2449 2587 2598 2611 2634 2694 2749 2778
#> [61] 2787 2842 2894 2909 3039 3045 3046 3074 3162 3174 3333 3352 3392 3403