Extract tips for a single site

The functions in sitePath usually include the results on more than one site. The function extractSite can be used to extract the predicted result on a single site.

extractSite(x, site, ...)

# S3 method for fixationSites
extractSite(x, site, ...)

Arguments

x
A fixationSites or a parallelSites object. More type will be supported in the later version.

site
A site included in the result.

...
Other arguments

Value

The predicted result of a single site

Examples

data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
mutations <- fixationSites(lineagePath(tree))
extractSite(mutations, 139)
#> Site 139 may experience fixation on 1 path(s):
#> 
#> S139N (20->72) 
#> 
#> In the bracket are the number of tips involved before and after the fixation