Extract info for fixation of a single site

The result of fixationSites contains all the possible sites with fixation mutation. The function extractTips retrieves the name of the tips involved in the fixation.

The function extractSite can be used to extract the fixation info of a single site.

# S3 method for fixationSites
extractTips(x, site, select = 1, ...)

# S3 method for multiFixationSites
extractTips(x, site, select = 1, ...)

# S3 method for sitePath
extractTips(x, select = 1, ...)

# S3 method for fixationSites
extractSite(x, site, ...)

# S3 method for multiFixationSites
extractSite(x, site, ...)

Arguments

x
A fixationSites or a sitePath object.

site
A site predicted to experience fixation.

select
For a site, there theoretically might be more than one fixation on different lineages. You may use this argument to extract for a specific fixation of a site. The default is the first fixation of the site.

...
Other arguments

Value

The function extractTips returns the name of the tips involved in the fixation.

The function extractSite returns a sitePath object

Examples

data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
mutations <- fixationSites(lineagePath(tree))
extractTips(mutations, 139)

#> [[1]]
#>  [1] "ANK57896" "AMD61711" "AQS26698" "APG56458" "AUI42289" "AMR39834"
#>  [7] "AWH65848" "APO08504" "AMX81917" "AVZ47169" "AMX81916" "AMD61710"
#> [13] "AMK49492" "AMX81915" "AOC50652" "APH11611" "BBC70847" "AUF35022"
#> [19] "ATL14618" "AUF35021"
#> attr(,"AA")
#> [1] "S"
#> 
#> [[2]]
#>  [1] "ANO46307" "AVG19275" "ANN44857" "ANO46302" "ANO46301" "AMD16557"
#>  [7] "APB03018" "AWH65849" "AMC13911" "APB03019" "ARB07953" "ARU07076"
#> [13] "ARU07074" "AMX81919" "AMB37295" "ANC90426" "ALU33341" "ASB32509"
#> [19] "AMH87239" "AMA12086" "AMA12085" "AMU04506" "AMA12087" "AMA12084"
#> [25] "AOY08538" "AOG18295" "AOY08517" "AOY08541" "AMM39805" "ARX97119"
#> [31] "AMK79468" "AML82110" "AMR39831" "AMX81918" "APY24200" "AOO19564"
#> [37] "AUI42194" "AOG18296" "ART29828" "AQZ41954" "AOY08533" "ATI21641"
#> [43] "ATG29291" "ATG29301" "AWF93617" "APC60215" "APO08503" "ATG29306"
#> [49] "ATG29284" "AQM74761" "ATG29287" "ATG29303" "AWF93619" "AWF93618"
#> [55] "AOW32303" "AMC13913" "AMC13912" "AOC50654" "ATG29292" "ATG29295"
#> [61] "AOY08536" "ASL68974" "AMQ48986" "ATG29315" "ATG29307" "ATG29270"
#> [67] "AQZ41952" "AQZ41955" "ARI68105" "ASU55505" "AQM74762" "AUD54964"
#> attr(,"AA")
#> [1] "N"
#> 

extractSite(mutations, 139)

#> Site 139 may experience fixation on 1 path(s):
#> 
#> S139N (20->72) 
#> 
#> In the bracket are the number of tips involved before and after the fixation