Accumulation of fixed mutation as a tree

The tips are clustered according to the fixation sites. The transition of fixation sites will be plotted as a phylogenetic tree. The length of each branch represents the number of fixation mutation between two clusters. The name of the tree tips indicate the number of sequences in the cluster.

fixationPath(x, ...)

# S3 method for sitesMinEntropy
fixationPath(x, minEffectiveSize = NULL, ...)

# S3 method for fixationSites
fixationPath(x, minEffectiveSize = NULL, ...)

Arguments

x
The return from fixationSites function.

...
Further arguments passed to or from other methods.

minEffectiveSize
The minimum size for a tip cluster.

Value

An fixationPath object

Examples

data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
paths <- lineagePath(tree)
mutations <- fixationSites(paths)
fixationPath(mutations)

#> Warning: `mutate_()` is deprecated as of dplyr 0.7.0.
#> Please use `mutate()` instead.
#> See vignette('programming') for more help
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> [1] "1.1.1" "2.1.1" "2.2.2" "2.2.4"