The grouping of tree tips

groupTips uses sequence similarity to group tree tips. Members in a group are always constrained to share the same ancestral node. Similarity between two tips is derived from their multiple sequence alignment. The site will not be counted into total length if both are gap. Similarity is calculated as number of matched divided by the corrected total length. So far the detection of divergence is based on one simple rule: the minimal pairwise similarity. The two branches are decided to be divergent if the similarity is lower than the threshold.

# S3 method for phyMSAmatched
groupTips(
  tree,
  similarity = NULL,
  simMatrix = NULL,
  forbidTrivial = TRUE,
  tipnames = TRUE,
  ...
)

# S3 method for sitesMinEntropy
groupTips(tree, tipnames = TRUE, ...)

# S3 method for fixationSites
groupTips(tree, tipnames = TRUE, ...)

# S3 method for fixationPath
groupTips(tree, tipnames = TRUE, ...)

Arguments

tree
The return from addMSA, lineagePath or other functions in the sitePath package.

similarity
Similarity threshold for tree trimming in groupTips. If not provided, the mean similarity subtract standard deviation of all sequences will be used.

simMatrix
A diagonal matrix of similarities for each pair of sequences.

forbidTrivial
Does not allow trivial trimming.

tipnames
If return tips as integer or tip names.

...
Other arguments.

Value

groupTips returns grouping of tips.

Examples

data(zikv_tree)
data(zikv_align)
tree <- addMSA(zikv_tree, alignment = zikv_align)
groupTips(tree)

#> $`122`
#> [1] "ANK57896"
#> 
#> $`504`
#>   [1] "APO39243"     "APO39229"     "ARU07183"     "AQZ41956"     "ATG29276"    
#>   [6] "ANG09399"     "AVZ25033"     "ASL68979"     "ATG29299"     "ASL68978"    
#>  [11] "AQZ41953"     "AQZ41949"     "ANO46297"     "ANO46298"     "AQX32985"    
#>  [16] "AQZ41947"     "AQZ41950"     "AQZ41951"     "AQZ41948"     "ATG29267"    
#>  [21] "AOY08518"     "ARB07962"     "AML81028"     "APY24199"     "ALX35659"    
#>  [26] "ANQ92019"     "AOX49264"     "AOX49265"     "APY24198"     "APO36913"    
#>  [31] "AQU12485"     "AMQ48981"     "APH11492"     "AOO54270"     "AND01116"    
#>  [36] "AMS00611"     "ANF04752"     "AOE22997"     "ASU55411"     "ASU55395"    
#>  [41] "ASU55396"     "ASU55392"     "AOY08524"     "ASU55394"     "ASU55404"    
#>  [46] "ASU55423"     "ANB66182"     "ASU55425"     "APQ41782"     "ASU55406"    
#>  [51] "ASU55398"     "ASU55401"     "ASU55402"     "ASU55407"     "AQX32987"    
#>  [56] "APB03017"     "ASU55399"     "APQ41783"     "ASU55410"     "ASU55397"    
#>  [61] "ANS60026"     "ASU55408"     "ASU55413"     "ASU55414"     "ASU55405"    
#>  [66] "AMZ03557"     "AMM39804"     "ANC90428"     "ASU55421"     "ASU55415"    
#>  [71] "APQ41786"     "ASU55390"     "AQX32986"     "ASU55422"     "ASU55426"    
#>  [76] "ASU55412"     "ASU55419"     "ARM59239"     "ANB66183"     "ASU55420"    
#>  [81] "ASU55418"     "ASU55424"     "APQ41784"     "ASU55391"     "ASU55400"    
#>  [86] "ASU55409"     "ANB66184"     "ASU55393"     "AMC33116"     "AMQ34003"    
#>  [91] "AMQ34004"     "ASU55403"     "ARB07967"     "AMK49165"     "ARB07976"    
#>  [96] "ASU55416"     "ANW07476"     "AMZ03556"     "ASU55417"     "ANK57897"    
#> [101] "AWH65849"     "AMC13911"     "APB03019"     "APB03018"     "AMD16557"    
#> [106] "ANO46302"     "ANO46301"     "AVG19275"     "ANN44857"     "ANO46307"    
#> [111] "BBC70847"     "AUF35022"     "ATL14618"     "AUF35021"     "APO08504"    
#> [116] "AMX81917"     "AVZ47169"     "AMX81916"     "AMR39834"     "AWH65848"    
#> [121] "AUI42289"     "AMD61710"     "AOC50652"     "APH11611"     "AMK49492"    
#> [126] "AMX81915"     "APH11550"     "APH11586"     "APH11535"     "APH11544"    
#> [131] "APH11584"     "APH11567"     "APH11548"     "APH11572"     "APH11575"    
#> [136] "APH11568"     "APH11561"     "APH11533"     "APH11523"     "AWK27349"    
#> [141] "APH11545"     "APH11554"     "APH11558"     "APH11564"     "APH11553"    
#> [146] "APH11555"     "APH11577"     "APH11525"     "APH11570"     "APH11565"    
#> [151] "APH11541"     "APH11546"     "APH11582"     "APH11585"     "APH11563"    
#> [156] "APH11504"     "APH11587"     "APH11499"     "APH11498"     "APH11503"    
#> [161] "APH11578"     "APH11506"     "APH11507"     "APH11497"     "APH11500"    
#> [166] "APH11502"     "AVG19202"     "APH11539"     "APH11534"     "APH11592"    
#> [171] "APH11594"     "APH11518"     "APH11549"     "APH11512"     "APH11514"    
#> [176] "APH11581"     "APH11508"     "APH11516"     "APH11515"     "APH11538"    
#> [181] "APH11551"     "APH11571"     "APH11566"     "APH11562"     "APH11595"    
#> [186] "APH11528"     "APH11560"     "APH11543"     "APH11529"     "APH11573"    
#> [191] "APH11536"     "APH11493"     "APH11532"     "APH11590"     "APH11559"    
#> [196] "APH11530"     "APH11517"     "APH11527"     "APH11593"     "APH11531"    
#> [201] "APH11576"     "APH11580"     "APH11569"     "APH11542"     "APH11579"    
#> [206] "APH11540"     "APH11557"     "APH11495"     "AOO19565"     "APH11511"    
#> [211] "APH11547"     "APH11524"     "APH11505"     "APH11501"     "APH11583"    
#> [216] "APH11537"     "APH11521"     "APH11509"     "APH11574"     "APH11552"    
#> [221] "AOL02459"     "APG56499"     "ASW34302"     "AVV62004"     "BAX00477"    
#> [226] "BAV89190"     "AOI20067"     "AMM43325"     "AMR68932"     "AMV49165"    
#> [231] "AMM39806"     "AMR39830"     "ANF16414"     "AMT75536"     "AMO03410"    
#> [236] "ANA12599"     "ANC90425"     "AUI42329"     "AUI42330"     "AMM43326"    
#> [241] "ANO46306"     "ANO46309"     "ANO46305"     "ANO46303"     "ARB08102"    
#> [246] "AHZ13508"     "ANO46304"     "ANO46308"     "ATB53752"     "AOR82893"    
#> [251] "ANH10698"     "AOR82892"     "ARU07075"     "AMQ48982"     "AMK49164"    
#> [256] "APG56457"     "AMB18850"     "YP_009428568" "ART29823"     "AVZ25035"    
#> [261] "AOY08542"     "ARB07932"     "AOS90223"     "AOS90220"     "APO39232"    
#> [266] "AOO85388"     "AOO53981"     "AOS90222"     "APO39228"     "AOY08546"    
#> [271] "AOY08521"     "APO39237"     "APB03020"     "AOY08516"     "APB03021"    
#> [276] "APO39236"     "AOS90221"     "APO39233"     "ASW34087"     "ART29826"    
#> [281] "ART29825"     "AOY08523"     "AOS90224"     "BAV82373"     "AOY08525"    
#> [286] "AOY08535"     "APW84875"     "APW84874"     "ANH22038"     "APW84877"    
#> [291] "APW84873"     "APW84872"     "AMN14620"     "APW84876"     "AOY08537"    
#> [296] "ASK51714"     "ARB07953"     "ARU07076"     "ARU07074"     "AMX81919"    
#> [301] "ANC90426"     "ALU33341"     "ASB32509"     "AMB37295"     "AMA12086"    
#> [306] "AMA12085"     "AMU04506"     "AMA12087"     "AMA12084"     "AMH87239"    
#> [311] "AOY08538"     "AOY08517"     "AOY08541"     "AOG18295"     "AMM39805"    
#> [316] "AMK79468"     "AML82110"     "AMR39831"     "AMX81918"     "ARX97119"    
#> [321] "APY24200"     "AUI42194"     "AOG18296"     "AOO19564"     "ATG29306"    
#> [326] "ATG29284"     "AQM74761"     "ATG29287"     "ATG29303"     "AWF93619"    
#> [331] "AWF93618"     "AWF93617"     "APC60215"     "ART29828"     "ATG29291"    
#> [336] "ATG29301"     "AOY08533"     "AQZ41954"     "ATI21641"     "APO08503"    
#> [341] "ATG29292"     "ATG29295"     "AOY08536"     "ASL68974"     "AMQ48986"    
#> [346] "ATG29315"     "ATG29307"     "ATG29270"     "ARI68105"     "ASU55505"    
#> [351] "AQM74762"     "AUD54964"     "AOW32303"     "AOC50654"     "AQZ41952"    
#> [356] "AQZ41955"     "AMC13913"     "AMC13912"    
#> 
#> $`507`
#> [1] "AMD61711" "AQS26698" "APG56458"
#>