Deprecated functions in package ‘sitePath’

These functions are provided for compatibility with older versions of ‘sitePath’ only, and will be defunct at the next release.

# S3 method for lineagePath
multiFixationSites(
  paths,
  samplingSize = NULL,
  samplingTimes = 100,
  minEffectiveSize = 0,
  searchDepth = 1,
  method = c("compare", "insert", "delete"),
  ...
)

Arguments

paths
a lineagePath object returned from lineagePath function or a phylo object after addMSA

samplingSize
The number of tips sampled for each round of resampling. It shoud be at least 10th and at most nine 10ths of the tree tips.

samplingTimes
The total times of random sampling to do. It should be greater than 100.

minEffectiveSize
A vector of two integers to specifiy minimum tree tips involved before/after mutation. Otherwise the mutation will not be counted into the return. If more than one number is given, the ancestral takes the first and descendant takes the second as the minimum. If only given one number, it's the minimum for both ancestral and descendant.

searchDepth
The function uses heuristic search but the termination of the search cannot be intrinsically decided. searchDepth is needed to tell the search when to stop.

method
The strategy for predicting the fixation. The basic approach is entropy minimization and can be achieved by adding or removing fixation point, or by comparing the two.

...
further arguments passed to or from other methods.

Value

Sites with multiple fixations.

Examples

data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
fixationSites(lineagePath(tree))

#> Result for 3 paths:
#> 
#> 123 196 763 988 1949 784 139 894 2074 2634 3045 
#> No reference sequence specified. Using alignment numbering