Topology-dependent tree trimming

similarityMatrix calculates similarity between aligned sequences. The similarity matrix can be used in groupTips.

groupTips uses sequence similarity to group tree tips. Members in a group are always constrained to share the same ancestral node. Similarity between two tips is derived from their multiple sequence alignment. The site will not be counted into total length if both are gap. Similarity is calculated as number of matched divided by the corrected total length. So far the detection of divergence is based on one simple rule: the miminal pairwise similarity. The two branches are decided to be divergent if the similarity is lower than the threshold.

similarityMatrix(tree)

groupTips(
  tree,
  similarity = NULL,
  simMatrix = NULL,
  forbidTrivial = TRUE,
  tipnames = TRUE
)

Arguments

tree
The return from addMSA function.

similarity
Similarity threshold for tree trimming in groupTips. If not provided, the mean similarity substract standard deviation of all sequences will be used.

simMatrix
A diagonal matrix of similarities for each pair of sequences.

forbidTrivial
Does not allow trivial trimming.

tipnames
If return tips as integer or tip names.

Value

similarityMatrix returns a diagonal matrix of similarity between sequences

groupTips returns grouping of tips.

Examples

data('zikv_tree')
data('zikv_align')
tree <- addMSA(zikv_tree, alignment = zikv_align)
simMatrix <- similarityMatrix(tree)
groupTips(tree, 0.996, simMatrix)

#> $`3`
#> [1] "ARU07183"
#> 
#> $`7`
#> [1] "AVZ25033"
#> 
#> $`20`
#> [1] "ATG29267"
#> 
#> $`24`
#> [1] "APY24199"
#> 
#> $`105`
#> [1] "AMD16557"
#> 
#> $`121`
#> [1] "AUI42289"
#> 
#> $`122`
#> [1] "ANK57896"
#> 
#> $`129`
#> [1] "AMK49492"
#> 
#> $`130`
#> [1] "AMX81915"
#> 
#> $`251`
#> [1] "ANO46304"
#> 
#> $`252`
#> [1] "ANO46308"
#> 
#> $`263`
#> [1] "ART29823"
#> 
#> $`302`
#> [1] "ARU07076"
#> 
#> $`325`
#> [1] "APY24200"
#> 
#> $`342`
#> [1] "AQZ41954"
#> 
#> $`343`
#> [1] "ATI21641"
#> 
#> $`344`
#> [1] "APO08503"
#> 
#> $`357`
#> [1] "AOW32303"
#> 
#> $`367`
#> [1] "APO39243" "APO39229" "AOC50654" "AQZ41952" "AQZ41955" "AMC13913" "AMC13912"
#> 
#> $`372`
#> [1] "AQZ41956" "ATG29276" "ANG09399"
#> 
#> $`380`
#> [1] "ASL68979" "ATG29299" "ASL68978" "AQZ41953" "AQZ41949" "ANO46297" "ANO46298"
#> 
#> $`387`
#> [1] "AQX32985" "AQZ41947" "AQZ41950" "AQZ41951" "AQZ41948"
#> 
#> $`397`
#> [1] "AOY08518" "ARB07962" "AML81028"
#> 
#> $`403`
#> [1] "ALX35659" "ANQ92019" "AOX49264" "AOX49265" "APY24198" "APO36913"
#> 
#> $`409`
#>  [1] "AQU12485" "AMQ48981" "APH11492" "AOO54270" "AND01116" "AMS00611"
#>  [7] "ANF04752" "AOE22997" "ASU55411" "ASU55395" "ASU55396" "ASU55392"
#> [13] "AOY08524" "ASU55394" "ASU55404" "ASU55423" "ANB66182" "ASU55425"
#> [19] "APQ41782" "ASU55406" "ASU55398" "ASU55401" "ASU55402" "ASU55407"
#> [25] "AQX32987" "APB03017" "ASU55399" "APQ41783" "ASU55410" "ASU55397"
#> [31] "ANS60026" "ASU55408" "ASU55413" "ASU55414" "ASU55405" "AMZ03557"
#> [37] "AMM39804" "ANC90428" "ASU55421" "ASU55415" "APQ41786" "ASU55390"
#> [43] "AQX32986" "ASU55422" "ASU55426" "ASU55412" "ASU55419" "ARM59239"
#> [49] "ANB66183" "ASU55420" "ASU55418" "ASU55424" "APQ41784" "ASU55391"
#> [55] "ASU55400" "ASU55409" "ANB66184" "ASU55393" "AMC33116" "AMQ34003"
#> [61] "AMQ34004" "ASU55403" "ARB07967"
#> 
#> $`472`
#>  [1] "AMK49165" "ARB07976" "ASU55416" "ANW07476" "AMZ03556" "ASU55417"
#>  [7] "ANK57897" "AWH65849" "AMC13911" "APB03019" "APB03018"
#> 
#> $`487`
#> [1] "ANO46302" "ANO46301"
#> 
#> $`490`
#> [1] "AVG19275" "ANN44857" "ANO46307"
#> 
#> $`496`
#> [1] "BBC70847" "AUF35022" "ATL14618" "AUF35021"
#> 
#> $`502`
#> [1] "APO08504" "AMX81917"
#> 
#> $`503`
#> [1] "AVZ47169" "AMX81916"
#> 
#> $`506`
#> [1] "AMR39834" "AWH65848"
#> 
#> $`507`
#> [1] "AMD61711" "AQS26698" "APG56458"
#> 
#> $`510`
#> [1] "AMD61710" "AOC50652" "APH11611"
#> 
#> $`513`
#>  [1] "APH11550" "APH11586" "APH11535" "APH11544" "APH11584" "APH11567"
#>  [7] "APH11548" "APH11572" "APH11575" "APH11568" "APH11561" "APH11533"
#> [13] "APH11523" "AWK27349" "APH11545" "APH11554" "APH11558" "APH11564"
#> [19] "APH11553" "APH11555" "APH11577" "APH11525" "APH11570" "APH11565"
#> [25] "APH11541" "APH11546" "APH11582" "APH11585" "APH11563" "APH11504"
#> [31] "APH11587" "APH11499" "APH11498" "APH11503" "APH11578" "APH11506"
#> [37] "APH11507" "APH11497" "APH11500" "APH11502" "AVG19202" "APH11539"
#> [43] "APH11534" "APH11592" "APH11594" "APH11518" "APH11549" "APH11512"
#> [49] "APH11514" "APH11581" "APH11508" "APH11516" "APH11515" "APH11538"
#> [55] "APH11551" "APH11571" "APH11566" "APH11562" "APH11595" "APH11528"
#> [61] "APH11560" "APH11543" "APH11529" "APH11573" "APH11536" "APH11493"
#> [67] "APH11532" "APH11590" "APH11559" "APH11530" "APH11517" "APH11527"
#> [73] "APH11593" "APH11531" "APH11576" "APH11580" "APH11569" "APH11542"
#> [79] "APH11579" "APH11540" "APH11557" "APH11495" "AOO19565" "APH11511"
#> [85] "APH11547" "APH11524" "APH11505" "APH11501" "APH11583" "APH11537"
#> [91] "APH11521" "APH11509" "APH11574" "APH11552" "AOL02459" "APG56499"
#> [97] "ASW34302"
#> 
#> $`609`
#> [1] "AVV62004" "BAX00477"
#> 
#> $`610`
#>  [1] "BAV89190" "AOI20067" "AMM43325" "AMR68932" "AMV49165" "AMM39806"
#>  [7] "AMR39830" "ANF16414" "AMT75536" "AMO03410" "ANA12599" "ANC90425"
#> [13] "AUI42329" "AUI42330" "AMM43326"
#> 
#> $`624`
#> [1] "ANO46306" "ANO46309" "ANO46305" "ANO46303"
#> 
#> $`627`
#> [1] "ARB08102" "AHZ13508"
#> 
#> $`631`
#> [1] "ATB53752" "AOR82893" "ANH10698" "AOR82892" "ARU07075" "AMQ48982" "AMK49164"
#> [8] "APG56457"
#> 
#> $`639`
#> [1] "AMB18850"     "YP_009428568"
#> 
#> $`641`
#>  [1] "AVZ25035" "AOY08542" "ARB07932" "AOS90223" "AOS90220" "APO39232"
#>  [7] "AOO85388" "AOO53981" "AOS90222" "APO39228" "AOY08546" "AOY08521"
#> [13] "APO39237" "APB03020" "AOY08516" "APB03021" "APO39236" "AOS90221"
#> [19] "APO39233" "ASW34087" "ART29826" "ART29825" "AOY08523" "AOS90224"
#> [25] "BAV82373" "AOY08525" "AOY08535" "APW84875" "APW84874" "ANH22038"
#> [31] "APW84877" "APW84873" "APW84872" "AMN14620" "APW84876" "AOY08537"
#> [37] "ASK51714" "ARB07953"
#> 
#> $`678`
#> [1] "ARU07074" "AMX81919" "ANC90426" "ALU33341" "ASB32509" "AMB37295"
#> 
#> $`683`
#> [1] "AMA12086" "AMA12085" "AMU04506" "AMA12087" "AMA12084" "AMH87239"
#> 
#> $`688`
#>  [1] "AOY08538" "AOY08517" "AOY08541" "AOG18295" "AMM39805" "AMK79468"
#>  [7] "AML82110" "AMR39831" "AMX81918" "ARX97119"
#> 
#> $`697`
#> [1] "AUI42194" "AOG18296" "AOO19564"
#> 
#> $`699`
#>  [1] "ATG29306" "ATG29284" "AQM74761" "ATG29287" "ATG29303" "AWF93619"
#>  [7] "AWF93618" "AWF93617" "APC60215" "ART29828"
#> 
#> $`708`
#> [1] "ATG29291" "ATG29301" "AOY08533"
#> 
#> $`710`
#> [1] "ATG29292" "ATG29295" "AOY08536" "ASL68974" "AMQ48986" "ATG29315" "ATG29307"
#> [8] "ATG29270"
#> 
#> $`719`
#> [1] "ARI68105" "ASU55505"
#> 
#> $`720`
#> [1] "AQM74762" "AUD54964"
#>