The fixation sites with mutation on parallel lineage

The operation between the results of fixationSites and parallelSites.

paraFixSites(x, ...)

# S3 method for lineagePath
paraFixSites(
  x,
  category = c("intersect", "union", "parallelOnly", "fixationOnly"),
  minEffectiveSize = NULL,
  searchDepth = 1,
  method = c("compare", "insert", "delete"),
  minSNP = NULL,
  mutMode = c("all", "exact", "pre", "post"),
  ...
)

# S3 method for sitesMinEntropy
paraFixSites(
  x,
  category = c("intersect", "union", "parallelOnly", "fixationOnly"),
  minSNP = NULL,
  mutMode = c("all", "exact", "pre", "post"),
  ...
)

Arguments

x
A lineagePath object returned from lineagePath function.

...
further arguments passed to or from other methods.

category
Could be parallelOnly, fixationOnly, intersect or union.

minEffectiveSize
The minimum number of tips in a group.

searchDepth
The function uses heuristic search but the termination of the search cannot be intrinsically decided. searchDepth is needed to tell the search when to stop.

method
The strategy for predicting the fixation. The basic approach is entropy minimization and can be achieved by adding or removing fixation point, or by comparing the two.

minSNP
The minimum number of mutations to be qualified as parallel on at least two lineages. The default is 1.

mutMode
The strategy for finding parallel site. The default all is to consider any mutation regardless of the amino acid/nucleotide before and after mutation; Or exact to force mutation to be the same; Or pre/post to select the site having amino acid/nucleotide before/after mutation.

Value

A paraFixSites object.

Examples

data(zikv_tree_reduced)
data(zikv_align_reduced)
tree <- addMSA(zikv_tree_reduced, alignment = zikv_align_reduced)
paraFixSites(lineagePath(tree))
#> No qualified sites found.